+++ src/main/java/org/bootcamp/javazoo/dto/SellerDto.java
@@ -6,8 +6,6 @@ import lombok.Getter;
 
 @Getter
 public class SellerDto extends UserDto{
-
-
     public SellerDto(Integer user_id, String user_name) {
         super(user_id, user_name);
     }
