--- src/main/java/org/bootcamp/javazoo/service/interfaces/IProductService.java
@@ -1,9 +1,0 @@
-package org.bootcamp.javazoo.service.interfaces;
-
-import org.bootcamp.javazoo.dto.ProductDto;
-import org.bootcamp.javazoo.entity.Product;
-
-public interface IProductService {
-    ProductDto mapToProductDto (Product productToMap);
-    Product convertDtoToProduct(ProductDto productDto);
-}
